backend/pyproject.toml
[project]
name = "arctic-base"
version = "0.1.0"
description = "Arctic Base — self-hosted workbench host for AI agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.18",
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "ruamel.yaml>=0.18",
    "pillow>=11.0",
    "python-ulid>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.7",
]

[project.scripts]
arctic-base = "arctic_base.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arctic_base"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
